refactor: Presentation 데이터 매핑 로직 분리 및 Repository 리팩터링
* **refactor: `PresentationMapper` 추가 및 데이터 변환 로직 이동**
    * `PresentationRepositoryImpl` 내부에 비대하게 존재하던 도메인 모델과 네트워크 DTO 간의 변환 로직을 별도의 `PresentationMapper.kt` 파일로 추출했습니다.
    * `PresentationAnalysisSummary`, `PresentationScriptDetail`, `PresentationWordDetail` 등 주요 데이터 구조에 대한 `toDomain` 확장 함수를 정의했습니다.
    * `Category`, `Purpose`, `Style`, `Audience` 등 도메인 모델과 서버 간의 타입 매핑 로직(enum to string)을 통합 관리하도록 개선했습니다.
    * `kotlinx-datetime`을 사용하여 날짜 문자열을 `Instant`로 파싱하는 로직(`toAnalysisInstant`)에 대한 예외 처리를 강화했습니다.

* **refactor: `PresentationRepositoryImpl` 구조 개선**
    * 데이터 매핑 로직이 분리됨에 따라 Repository 내의 불필요한 private 메서드들을 제거하고 코드를 간결화했습니다.
    * 네트워크 요청 시 `PresentationAnalysisRequest`를 `PresentationAnalysisRequestParts`로 변환할 때 새로운 매퍼를 사용하도록 수정했습니다.
